[embedding] polish parallel embedding tablewise (#1545)
